# Conflicts:
#	UserSettings/Layouts/default-2022.dwlt

+++ Assets/_Project/Runtime/Core/Health/Health.cs
@@ -6,24 +6,24 @@ namespace _Project.Runtime.Core.Health
 {
     public class Health : IHealth, IReadOnlyHealth
     {
-        private readonly ReactiveProperty<int> _value = new();
+        private readonly ReactiveProperty<float> _value = new();
 
         public Health(IHealthConfig config)
         {
             MaxValue = config.MaxValue;
             _value.Value = MaxValue;
         }
-        public ReadOnlyReactiveProperty<int> Value => _value;
+        public ReadOnlyReactiveProperty<float> Value => _value;
         public int MaxValue { get; }
 
-        public void Increase(int value)
+        public void Increase(float value)
         {
             if (value <= 0)
                 throw new ArgumentException("Value must be greater than 0", nameof(value));
 
             _value.Value = Math.Clamp(_value.Value + value, 0, MaxValue);
         }
-        public void Decrease(int value)
+        public void Decrease(float value)
         {
             if (value <= 0)
                 throw new ArgumentException("Value must be greater than 0", nameof(value));
